Clean it in full
