+++ src/lib.rs
@@ -27,3 +27,6 @@ pub use probe::*;
 
 mod health;
 pub use health::*;
+
+mod engine_api;
+pub use engine_api::*;
